Add support for circular references
Support for circular references basically means
doing the same as with normal referential equality
but cutting off the JSON tree at some point.
That's what we're doing here! Instead of throwing `TypeError: Circular Reference`,
we just stop the recursion and return a `null`.
That's basically all ^^

To be honest, this is not super-intuitive for `JSON.stringify` readers
because the `null` can occur in kind-of unintuitive places.
But it's the best we can do - and since `JSON.stringify` doesn't even support
anything like this at all, I guess it's fine.