Transform undefined => null
Transformers shall always go from a JSON-incompatible
value to a JSON-compatible one.
Before, the `undefined` transformer
didn't change anything about the value,
so JSON.stringify had an effect: it changed undefined => null.
This commit fixes the transformer, so that no surprises happen
during serialisation.

While this changes the behaviour of `.serialize`,
this change is not considered breaking b/c it behaves the same
in the context of serialisation.