perf: require Node.js 20+, Safari 14+
This commit makes some backward compatibility-breaking changes to
improve performance:

1. Using `structuredClone` (available in Node.js 18+) instead of the
   `copy-anything` dependency.
2. Assuming `BigInt` (available in Safari 14+ and every other runtime
   from 2021 onward) always exists.

It also drops the ESLint rules that forbade various ES6 features. The
only reason not to use ES6 was IE11, which Microsoft officially ended
support for in 2022.

Since Node.js 18 reached its EOL date in April 2025, I've bumped the
`engines` declaration to `"node": ">=20"`. None of the changes here
require Node.js 20, but I think it's a good idea to make that bump in
case we want to take advantage of Node.js 20-specific features in the
future.